+++ HomeWork12/Sources/Extension/Constants.swift
@@ -1,0 +1,19 @@
+//
+//  Constants.swift
+//  HomeWork12
+//
+//  Created by Ренат Сафин on 09.02.2022.
+//
+
+import UIKit
+
+extension Metric {
+    static let tableViewRowHeight: CGFloat = 55
+    static let toSwitchTrailing: CGFloat = -20
+    static let notificationLabelWidth: CGFloat = 10
+    static let notificationLabelHeight: CGFloat = 10
+    static let notificationLabelContainerWhidth: CGFloat = 30
+    static let notificationLabelContainerHeight: CGFloat = 30
+    static let notificationLabelContainerTrailing: CGFloat = -20
+    static let viewCornerRadius: CGFloat = 15
+}
