+++ HomeWork12/Sources/Extension/TableViewDelegate.swift
@@ -9,11 +9,11 @@ import UIKit
 
 // MARK: - UITableViewDelegate
 
-extension ViewController: UITableViewDelegate {
+extension SettingView: UITableViewDelegate {
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
-        let model = data[indexPath.section][indexPath.row]
+        let model = ViewController.data[indexPath.section][indexPath.row]
         
         switch model.typeCell {
         case .cell:
