+++ HomeWork12/Sources/Scenes/ViewController.swift
@@ -8,64 +8,34 @@ import UIKit
 
 class ViewController: UIViewController {
     
-    var data = [[SettingsCell]]()
+    static var data = [[SettingsCell]]()
     
-    // MARK: - Views
-    
-    lazy var tableView: UITableView = {
-        let tableView = UITableView(frame: .zero, style: .grouped)
-        tableView.register(Cell.self, forCellReuseIdentifier: Cell.identifireCell)
-        tableView.register(SwitchCell.self, forCellReuseIdentifier: SwitchCell.identifireSwitchCell)
-        tableView.register(NotificationCell.self, forCellReuseIdentifier: NotificationCell.identifierNotificationCell)
-        tableView.rowHeight = Metric.tableViewRowHeight
-        tableView.dataSource = self
-        tableView.delegate = self
-        return tableView
-    }()
+    private var onboardinView: SettingView? {
+        guard isViewLoaded else { return nil }
+        return view as? SettingView
+    }
     
     // MARK: - Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupHierarhy()
-        setupLayout()
+        view = SettingView()
         setupViewNavigationBar()
-        updateData(addTo: FirstSection.addCellsToSection())
-        updateData(addTo: SecondSection.addCellsToSection())
-        updateData(addTo: ThirdSection.addCellsToSection())
-    }
-    
-    // MARK: - Settings
-    
-    private func setupHierarhy() {
-        view.addSubview(tableView)
-    }
-    
-    private func setupLayout() {
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor).isActive = true
-        tableView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor).isActive = true
-        tableView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor).isActive = true
-        tableView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor).isActive = true
+        updateData(addTo: SettingsModel.addCellsToSections())
     }
     
     // MARK: - SettingsNavigationBar
     
     private func setupViewNavigationBar() {
-        view.backgroundColor = .white
         title = "Настройки"
     }
     
     // MARK: - UpdateData
     
     private func updateData(addTo: [[SettingsCell]]) -> [[SettingsCell]] {
-        data.append(contentsOf: addTo)
-        return data
+        ViewController.data.append(contentsOf: addTo)
+        return ViewController.data
     }
 }
 
-// MARK: - Constants
 
-extension Metric {
-    static let tableViewRowHeight: CGFloat = 55
-}
