+++ HomeWork12/Sources/Scenes/Views/ViewCell.swift
@@ -1,27 +1,12 @@
 //
-//  Cell.swift
+//  ViewCell.swift
 //  HomeWork12
 //
 //  Created by Ренат Сафин on 07.01.2022.
 //
 import UIKit
 
-struct SettingsCell {
-    let nameImage: String
-    let colorImage: UIColor
-    let textLabel: String
-    let typeCell: TypeCell
-    let trailingLabelText: String?
-    let notificationLabelText: String?
-}
-
-enum TypeCell {
-    case cell
-    case switchCell
-    case notificationCell
-}
-
-class Cell: UITableViewCell {
+class ViewCell: UITableViewCell {
     
     static let identifireCell = "Cell"
